pyproject.toml
[tool.poetry]
name = "velluto-autopilot"
version = "0.1.0"
description = "Autonomous ecommerce growth OS for Velluto cycling eyewear"
authors = ["Velluto <tech@velluto.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { extras = ["standard"], version = "^0.32.0" }
anthropic = "^0.40.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }
asyncpg = "^0.30.0"
alembic = "^1.13.0"
redis = { extras = ["hiredis"], version = "^5.2.0" }
rq = "^1.16.0"
apscheduler = { extras = ["asyncio"], version = "^3.10.0" }
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
httpx = "^0.27.0"
structlog = "^24.4.0"
python-dotenv = "^1.0.0"
shopifyapi = "^12.6.0"
facebook-business = "^21.0.0"
google-auth = "^2.35.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.151.0"
sendgrid = "^6.11.0"
jinja2 = "^3.1.0"
pillow = "^11.0.0"
openai = "^1.54.0"
tenacity = "^9.0.0"
prometheus-client = "^0.21.0"
sentry-sdk = { extras = ["fastapi"], version = "^2.17.0" }
arrow = "^1.3.0"
beautifulsoup4 = "^4.12.0"
aiohttp = "^3.10.0"
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
respx = "^0.21.0"
freezegun = "^1.5.0"
factory-boy = "^3.3.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
